\name{CDMLearning-package}
\alias{CDMLearning-package}
\alias{CDMLearning}
\docType{package}
\title{
\packageTitle{CDMLearning}
}
\description{
\packageDescription{CDMLearning}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CDMLearning}
\packageIndices{CDMLearning}
}
\author{
\strong{Maintainer:} \packageMaintainer{CDMLearning}


Authors:
\itemize{
  \item Susu Zhang \email{susu.zhang1992@gmail.com}
  \item Shiyu Wang \email{swang44@uga.edu}
  \item Yinghan Chen \email{yinghanc@unr.edu}

  }
}
\references{
Wang, S., Yang, Y., Culpepper, S. A., & Douglas, J. A. (2018). Tracking Skill Acquisition With Cognitive Diagnosis Models: A Higher-Order, Hidden Markov Model With Covariates. Journal of Educational and Behavioral Statistics, 1076998617719727.

Chen, Y., Culpepper, S. A., Wang, S., & Douglas, J. (2018). A hidden Markov model for learning trajectories in cognitive diagnosis with application to spatial rotation skills. Applied Psychological Measurement, 42(1), 5-23.
}
\keyword{ package }
