% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_rotation_data.R
\docType{data}
\name{Design_array}
\alias{Design_array}
\title{Test design array}
\format{
An N-by-J-by-T array.
}
\source{
Spatial Rotation Learning Experiment at UIUC between Fall 2015 and Spring 2016.
}
\usage{
Design_array
}
\description{
This data set contains the design matrix indicating the items exposed to the subjects at each time point.
}
\author{
Shiyu Wang, Yan Yang, Jeff Douglas, and Steve Culpepper
}
\keyword{datasets}
