% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ETAmat}
\alias{ETAmat}
\title{Generate ideal response matrix}
\usage{
ETAmat(K, J, Q)
}
\arguments{
\item{K}{An \code{int} of the number of attributes}

\item{J}{An \code{int} of the number of items}

\item{Q}{A J-by-K Q \code{matrix}}
}
\value{
A J-by-2^K ideal response \code{matrix}
}
\description{
Based on the Q matrix and the latent attribute space, generate the ideal response matrix for each skill pattern
}
\examples{
Q = random_Q(15,4)
ETA = ETAmat(4,15,Q)
}
