% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmcdm-package.R
\docType{package}
\name{hmcdm-package}
\alias{hmcdm-package}
\alias{_PACKAGE}
\title{hmcdm: Hidden Markov Cognitive Diagnosis Models for Learning}
\description{
Functions for fitting hidden Markov models of learning under the cognitive diagnosis framework. This package enables the estimation of the hidden Markov diagnostic classification model, the first order hidden Markov model, the reduced-reparameterized unified learning model, and the joint learning model for responses and response times.
}
\references{
Wang, S., Yang, Y., Culpepper, S. A., & Douglas, J. A. (2018). Tracking Skill Acquisition With Cognitive Diagnosis Models: A Higher-Order, Hidden Markov Model With Covariates. Journal of Educational and Behavioral Statistics, 1076998617719727.

Chen, Y., Culpepper, S. A., Wang, S., & Douglas, J. (2018). A hidden Markov model for learning trajectories in cognitive diagnosis with application to spatial rotation skills. Applied Psychological Measurement, 42(1), 5-23.

Wang, S., Zhang, S., Douglas, J., & Culpepper, S. (2018). Using Response Times to Assess Learning Progress: A Joint Model for Responses and Response Times. Measurement: Interdisciplinary Research and Perspectives, 16(1), 45-58.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tmsalab/hmcdm}
  \item Report bugs at \url{https://github.com/tmsalab/hmcdm/issues}
}

}
\author{
\strong{Maintainer}: Sunbeom Kwon \email{sunbeom2@illinois.edu}

Authors:
\itemize{
  \item Susu Zhang \email{szhan105@illinois.edu}
  \item Shiyu Wang \email{swang44@uga.edu}
  \item Yinghan Chen \email{yinghanc@unr.edu }
}

}
