% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inv_bijectionvector}
\alias{inv_bijectionvector}
\title{Convert integer to attribute pattern}
\usage{
inv_bijectionvector(K, CL)
}
\arguments{
\item{K}{An \code{int} for the number of attributes}

\item{CL}{An \code{int} between 0 and 2^K-1}
}
\value{
A \code{vec} of the K-dimensional attribute pattern corresponding to CL.
}
\description{
Based on the bijective relationship between natural numbers and sum of powers of two,
convert integer between 0 and 2^K-1 to K-dimensional attribute pattern.
}
\examples{
inv_bijectionvector(4,0)
}
