% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_rotation_data.R
\docType{data}
\name{L_real_array}
\alias{L_real_array}
\title{Observed response times array}
\format{
An array of dimensions N-by-J-by-L. Each slice of the array is an N-by-J matrix, containing the
subjects' response times in seconds to each item at time point l.
}
\source{
Spatial Rotation Learning Experiment at UIUC between Fall 2015 and Spring 2016.
}
\usage{
L_real_array
}
\description{
\code{L_real_array} contains the observed latencies of responses of all subjects to all questions in the Spatial Rotation
Learning Program.
}
\author{
Shiyu Wang, Yan Yang, Jeff Douglas, and Steve Culpepper
}
\keyword{datasets}
