% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{List2Array}
\alias{List2Array}
\title{List to Array}
\usage{
List2Array(Q_List)
}
\arguments{
\item{Q_List}{A list of Q matrices}
}
\value{
Q_Array An array of Q matrices
}
\description{
Takes a list of (J*K) Q matrices as an input and returns an
array of Q matrices as an output.
}
\examples{
\donttest{
List2Array(Q_List)
}
}
