% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Mat2Array}
\alias{Mat2Array}
\title{Convert a J-by-K matrix to a Jt-by-K-by-T array}
\usage{
Mat2Array(Q_matrix, T)
}
\arguments{
\item{Q_matrix}{A J*K Q_matrix of the test}

\item{T}{The number of time points}
}
\value{
A Jt-by-K-by-T array of Q_matrices
}
\description{
Obtain an Jt-by-K-by-T array of Q-matrices each slice containing a Jt*K Q_matrix
}
\examples{
\donttest{
T = 5
Mat2Array(Q_matrix, T)}
}
