% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{point_estimates_learning}
\alias{point_estimates_learning}
\title{Obtain learning model point estimates}
\usage{
point_estimates_learning(output, model, N, Jt, K, T, alpha_EAP = TRUE)
}
\arguments{
\item{output}{A \code{list} of MCMC outputs, obtained from the MCMC_learning function}

\item{model}{A \code{charactor} of the type of model fitted with the MCMC sampler, possible selections are
"DINA_HO": Higher-Order Hidden Markov Diagnostic Classification Model with DINA responses;
"DINA_HO_RT_joint": Higher-Order Hidden Markov DCM with DINA responses, log-Normal response times, and joint modeling of latent
speed and learning ability;
"DINA_HO_RT_sep": Higher-Order Hidden Markov DCM with DINA responses, log-Normal response times, and separate modeling of latent
speed and learning ability;
"rRUM_indept": Simple independent transition probability model with rRUM responses
"NIDA_indept": Simple independent transition probability model with NIDA responses
"DINA_FOHM": First Order Hidden Markov model with DINA responses}

\item{N}{An \code{int} of number of subjects}

\item{Jt}{An \code{int} of number of items in each block}

\item{K}{An \code{int} of number of skills}

\item{T}{An \code{int} of number of time points}

\item{alpha_EAP}{A \code{boolean} operator (T/F) of whether to use EAP for alphas (if F: use most likely trajectory (MAP) for alphas)}
}
\value{
A \code{list} of point estimates of model parameters
}
\description{
Obtain EAPs of continuous parameters and EAP or MAP of the attribute trajectory estimates under
the CDM learning models based on the MCMC output
}
\examples{
\donttest{
N = length(Test_versions)
Jt = nrow(Q_list[[1]])
K = ncol(Q_list[[1]])
T = nrow(test_order)
J = Jt*T
output_FOHM = MCMC_learning(Y_real_list,Q_list,"DINA_FOHM",test_order,Test_versions,10000,5000)
point_estimates = point_estimates_learning(output_FOHM,"DINA_FOHM",N,Jt,K,T,alpha_EAP = T)
}
}
\author{
Susu Zhang
}
