% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_check.R
\name{pp_check.hmcdm}
\alias{pp_check.hmcdm}
\title{Graphical posterior predictive checks for hidden Markov cognitive diagnosis model}
\usage{
\method{pp_check}{hmcdm}(object, plotfun = "dens_overlay", type = "total_score")
}
\arguments{
\item{object}{a fitted model object of class "\code{hmcdm}".}

\item{plotfun}{A character string naming the type of plot. The list of available
plot functions include \code{"dens_overlay"}, \code{"hist"}, \code{"stat_2d"}, \code{"scatter_avg"}, \code{"error_scatter_avg"}.
The default function is \code{"dens_overlay"}.}

\item{type}{A character string naming the type of posterior predictive distributions to plot.
The list of available types include \code{"total_score"}, \code{"M1"}, \code{"M2"}. The default type is \code{"total_score"}.}
}
\description{
\code{pp_check} method for class \code{hmcdm}.
}
\examples{
\donttest{
output_FOHM = hmcdm(Y_real_array,Q_matrix,"DINA_FOHM",Test_order,Test_versions,10000,5000)
library(bayesplot)
pp_check(output_FOHM)
pp_check(output_FOHM, plotfun="hist", type="M1")
}
}
\references{
Zhang, S., Douglas, J. A., Wang, S. & Culpepper, S. A., 2019, Handbook of Diagnostic Classification Models: Models and Model Extensions, Applications, Software Packages. von Davier, M. & Lee, Y-S. (eds.). Springer, p. 503-524
}
\seealso{
\code{\link[bayesplot:PPC-distributions]{bayesplot::ppc_dens_overlay()}}
\code{\link[bayesplot:PPC-test-statistics]{bayesplot::ppc_stat()}}
\code{\link[bayesplot:PPC-test-statistics]{bayesplot::ppc_stat_2d()}}
\code{\link[bayesplot:PPC-scatterplots]{bayesplot::ppc_scatter_avg()}}
\code{\link[bayesplot:PPC-errors]{bayesplot::ppc_error_scatter_avg()}}
}
