% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_rotation_data.R
\docType{data}
\name{Q_examinee}
\alias{Q_examinee}
\title{List of Q-matrices for each examinee.}
\format{
A list of length 350. Each element of the list is a 50x4 matrix, containing the Q matrix of all items
administered across all time points to the examinee, in the order of administration.
}
\source{
Spatial Rotation Learning Experiment at UIUC between Fall 2015 and Spring 2016.
}
\usage{
Q_examinee
}
\description{
This data set contains the Q matrices for each subject in the Spatial Rotation Learning Program.
}
\author{
Shiyu Wang, Yan Yang, Jeff Douglas, and Steve Culpepper
}
\keyword{datasets}
