% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_rotation_data.R
\docType{data}
\name{Q_list}
\alias{Q_list}
\title{List of Q matrices}
\format{
A list of length 5 (number of item blocks). Each element of the list is a Jt-by-K matrix, containing the
item-skill relationship of items in the corresponding block.
}
\source{
Spatial Rotation Learning Experiment at UIUC between Fall 2015 and Spring 2016.
}
\usage{
Q_list
}
\description{
This data set contains the Q matrices of the items in the Spatial Rotation Learning Program.
}
\author{
Shiyu Wang, Yan Yang, Jeff Douglas, and Steve Culpepper
}
\keyword{datasets}
