% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Q_list_g}
\alias{Q_list_g}
\title{Generate a list of Q-matrices for each examinee.}
\usage{
Q_list_g(Q_matrix, Design_array)
}
\arguments{
\item{Q_matrix}{A J-by-K matrix, indicating the item-skill relationship.}

\item{Design_array}{An N-by-J-by-L array indicating whether examinee n has taken item j at l time point.}
}
\value{
A list length of N. Each element of the list is a JxK Q_matrix for each examinee.
}
\description{
Generate a list of length N. Each element of the list is a JxK Q_matrix of all items
administered across all time points to the examinee, in the order of administration.
}
\examples{
\donttest{
Q_examinee = Q_list_g(Q_matrix, Design_array)}
}
