% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_rotation_data.R
\docType{data}
\name{Q_matrix}
\alias{Q_matrix}
\title{Q-matrix}
\format{
A J-by-K matrix, indicating the item-skill relationship.
}
\source{
Spatial Rotation Learning Experiment at UIUC between Fall 2015 and Spring 2016.
}
\usage{
Q_matrix
}
\description{
\code{Q_matrix} contains the Q matrix of the items in the Spatial Rotation Learning Program.
}
\author{
Shiyu Wang, Yan Yang, Jeff Douglas, and Steve Culpepper
}
\keyword{datasets}
