% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_rotation_data.R
\docType{data}
\name{Qs}
\alias{Qs}
\title{Array of Q matrices}
\format{
An array of dimensions 10-by-4-by-5. Each slice of the array is a Jt-by-K matrix, containing the
item-skill relationship of items in the corresponding block.
}
\source{
Spatial Rotation Learning Experiment at UIUC between Fall 2015 and Spring 2016.
}
\usage{
Qs
}
\description{
This array contains the Q matrices of the items in the Spatial Rotation Learning Program.
}
\author{
Shiyu Wang, Yan Yang, Jeff Douglas, and Steve Culpepper
}
\keyword{datasets}
