% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rinvwish}
\alias{rinvwish}
\title{Generate Random Inverse Wishart Distribution}
\usage{
rinvwish(df, Sig)
}
\arguments{
\item{df}{An \code{int} that represents the degrees of freedom.  (> 0)}

\item{Sig}{A \code{matrix} with dimensions m x m that provides Sigma, the covariance matrix.}
}
\value{
A \code{matrix} that is an inverse wishart distribution.
}
\description{
Creates a random inverse wishart distribution when given degrees of freedom and a sigma matrix.
}
\examples{
#Call with the following data:
riwishart(3, diag(2))
}
\author{
James J Balamuta
}
