% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rOmega}
\alias{rOmega}
\title{Generate a random transition matrix for the first order hidden Markov model}
\usage{
rOmega(TP)
}
\arguments{
\item{TP}{A 2^K-by-2^K dichotomous matrix of indicating possible transitions under the monotonicity assumption, created with
the TPmat function}
}
\description{
Generate a random transition matrix under nondecreasing learning trajectory assumption
}
\examples{
N = length(Test_versions)
J = nrow(Q_matrix)
K = ncol(Q_matrix)
T = nrow(Test_order)
Jt = J/T
TP = TPmat(K)
Omega_sim = rOmega(TP)
}
