% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rOmegaFree}
\alias{rOmegaFree}
\title{Generate a random transition matrix for the first order hidden Markov model}
\usage{
rOmegaFree(TP)
}
\arguments{
\item{TP}{A 2^K-by-2^K dichotomous matrix of indicating possible transitions}
}
\description{
Generate a random transition matrix under the unrestricted learning trajectory assumption
}
\examples{
K = ncol(Q_list[[1]])
TP = TPmatFree(K)
Omega_sim = rOmega(TP)
}
