% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_resp_DINA}
\alias{sim_resp_DINA}
\title{Simulate DINA model responses (single vector)}
\usage{
sim_resp_DINA(J, K, ETA, Svec, Gvec, alpha)
}
\arguments{
\item{J}{An \code{int} of number of items}

\item{K}{An \code{int} of number of attributes}

\item{ETA}{A \code{matrix} of ideal responses generated with ETAmat function}

\item{Svec}{A length J \code{vector} of item slipping parameters}

\item{Gvec}{A length J \code{vector} of item guessing parameters}

\item{alpha}{A length K \code{vector} of attribute pattern of a person}
}
\value{
A length J \code{vector} of item responses
}
\description{
Simulate a single vector of DINA responses for a person on a set of items
}
\examples{
J = 15
K = 4
Q = random_Q(J,K)
ETA = ETAmat(K,J,Q)
s = runif(J,.1,.2)
g = runif(J,.1,.2)
alpha_i = c(1,0,0,1)
Y_i = sim_resp_DINA(J,K,ETA,s,g,alpha_i)
}
