% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_resp_NIDA}
\alias{sim_resp_NIDA}
\title{Simulate NIDA model responses (single vector)}
\usage{
sim_resp_NIDA(J, K, Q, Svec, Gvec, alpha)
}
\arguments{
\item{J}{An \code{int} of number of items}

\item{K}{An \code{int} of number of attributes}

\item{Q}{A J-by-K Q \code{matrix}}

\item{Svec}{A length K \code{vector} of slipping probability in applying mastered skills}

\item{Gvec}{A length K \code{vector} of guessing probability in applying mastered skills}

\item{alpha}{A length K \code{vector} of attribute pattern of a person}
}
\value{
A length J \code{vector} of item responses
}
\description{
Simulate a single vector of NIDA responses for a person on a set of items
}
\examples{
J = 15
K = 4
Q = random_Q(J,K)
Svec <- runif(K,.1,.3)
Gvec <- runif(K,.1,.3)
alpha_i = c(1,0,0,1)
Y_i = sim_resp_NIDA(J,K,Q,Svec,Gvec,alpha_i)
}
