% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_resp_rRUM}
\alias{sim_resp_rRUM}
\title{Simulate rRUM model responses (single vector)}
\usage{
sim_resp_rRUM(J, K, Q, rstar, pistar, alpha)
}
\arguments{
\item{J}{An \code{int} of number of items}

\item{K}{An \code{int} of number of attributes}

\item{Q}{A J-by-K Q \code{matrix}}

\item{rstar}{A J-by-K \code{matrix} of item penalty parameters for missing requisite skills}

\item{pistar}{length J \code{vector} of item correct response probability with all requisite skills}

\item{alpha}{A length K \code{vector} of attribute pattern of a person}
}
\value{
A length J \code{vector} of item responses
}
\description{
Simulate a single vector of rRUM responses for a person on a set of items
}
\examples{
J = 10
K = 4
Q = random_Q(J,K)
Smats <- matrix(runif(Jt*K,.1,.3),Jt,K)
Gmats <- matrix(runif(Jt*K,.1,.3),Jt,K)
r_stars <- matrix(NA,Jt,K)
pi_stars <- numeric(Jt)
for(t in 1:T){
  pi_stars <- apply(((1-Smats)^Q),1,prod)
  r_stars <- Gmats/(1-Smats)
}
alpha_i = c(1,0,0,1)
Y_i = sim_resp_rRUM(J,K,Q,r_stars,pi_stars,alpha_i)
}
