% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simDINA}
\alias{simDINA}
\title{Simulate DINA model responses (entire cube)}
\usage{
simDINA(alphas, itempars, ETA, test_order, Test_versions)
}
\arguments{
\item{alphas}{An N-by-K-by-T \code{array} of attribute patterns of all persons across T time points}

\item{itempars}{A J-by-2-by-T \code{cube} of item parameters (slipping: 1st col, guessin: 2nd col) across item blocks}

\item{ETA}{A J-by-2^K-by-T \code{array} of ideal responses across all item blocks, with each slice generated with ETAmat function}

\item{test_order}{A N_versions-by-T \code{matrix} indicating which block of items were administered to examinees with specific test version.}

\item{Test_versions}{A length N \code{vector} of the test version of each examinee}
}
\value{
An \code{array} of DINA item responses of examinees across all time points
}
\description{
Simulate a cube of DINA responses for all persons on items across all time points
}
\examples{
itempars_true <- array(runif(Jt*2*T,.1,.2), dim = c(Jt,2,T))

ETAs <- array(NA,dim = c(Jt,2^K,T)) 
for(t in 1:T){
  ETAs[,,t] <- ETAmat(K,Jt,Q_list[[t]])
}

Y_sim <- simDINA(Alphas,itempars_true,ETAs,test_order,Test_versions)
}
