% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sparse2Dense}
\alias{Sparse2Dense}
\title{Sparse2Dense}
\usage{
Sparse2Dense(Y_real_array, test_order, Test_versions)
}
\arguments{
\item{Y_real_array}{A N\emph{J}T sparse response array}

\item{test_order}{A test_order matrix}

\item{Test_versions}{A vector of test versions}
}
\value{
Response An array of response
}
\description{
Converts a N\emph{J}T sparse response array to a N\emph{Jt}T dense array.
}
\examples{
\donttest{
Sparse2Dense(Y_real_array, test_order, Test_versions)
}
}
