% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_rotation_data.R
\docType{data}
\name{test_order}
\alias{test_order}
\title{Test block ordering of each test version}
\format{A 5x5 matrix, each row is the order of item blocks (as in Qs and Q_list) for that test version.
For example, the first row is the order of item block administration (1-2-3-4-5) to subjects with test
version 1.}
\source{
Spatial Rotation Learning Experiment at UIUC between Fall 2015 and Spring 2016.
}
\usage{
test_order
}
\description{
This data set contains the item block ordering of each version of the test.
}
\author{
Shiyu Wang, Yan Yang, Jeff Douglas, and Steve Culpepper
}
\keyword{datasets}
