% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_rotation_data.R
\docType{data}
\name{Test_versions}
\alias{Test_versions}
\title{Subjects' test version}
\format{
A vector of length 350, containing each subject's test version ranging from 1 to 5.
}
\source{
Spatial Rotation Learning Experiment at UIUC between Fall 2015 and Spring 2016.
}
\usage{
Test_versions
}
\description{
This data set contains each subject's test version in the Spatial Rotation Learning Program.
}
\author{
Shiyu Wang, Yan Yang, Jeff Douglas, and Steve Culpepper
}
\keyword{datasets}
