% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TPmat}
\alias{TPmat}
\title{Generate monotonicity matrix}
\usage{
TPmat(K)
}
\arguments{
\item{K}{An \code{int} of the number of attributes.}
}
\value{
A 2^K-by-2^K dichotomous \code{matrix} of whether it is possible to transition between two patterns
}
\description{
Based on the latent attribute space, generate a matrix indicating whether it is possible to
transition from pattern cc to cc' under the monotonicity learning assumption.
}
\examples{
TP = TPmat(4)
}
