% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TPmatFree}
\alias{TPmatFree}
\title{Generate non-monotonicity matrix}
\usage{
TPmatFree(K)
}
\arguments{
\item{K}{An \code{int} of the number of attribtues.}
}
\value{
A 2^K-by-2^K \code{matrix} filled with 1, showing it is possible to transit between any two patterns
}
\description{
Based on the latent attribute space, generate a matrix indicating whether it is possible to
transition from pattern cc to cc', free from the monotonicity learning assumption.
}
\examples{
TP = TPmatFree(4)
}
