% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_rotation_data.R
\docType{data}
\name{Y_real_array}
\alias{Y_real_array}
\title{Observed response accuracy array}
\format{
An array of dimensions N-by-J-by-L. Each slice of the array is an N-by-J matrix, containing the
subjects' response accuracy to each item at time point l.
}
\source{
Spatial Rotation Learning Experiment at UIUC between Fall 2015 and Spring 2016.
}
\usage{
Y_real_array
}
\description{
\code{Y_real_array} contains each subject's observed response accuracy (0/1) at all time points in the Spatial
Rotation Learning Program.
}
\author{
Shiyu Wang, Yan Yang, Jeff Douglas, and Steve Culpepper
}
\keyword{datasets}
