% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_rotation_data.R
\docType{data}
\name{Y_real_list}
\alias{Y_real_list}
\title{Observed response accuracy list}
\format{
A list of length 5 (number of time points). Each element of the list is an N-by-Jt matrix, containing the
subjects' response accuracy to each item at that time point.
}
\source{
Spatial Rotation Learning Experiment at UIUC between Fall 2015 and Spring 2016.
}
\usage{
Y_real_list
}
\description{
This data set contains each subject's observed response accuracy (0/1) at all time points in the Spatial
Rotation Learning Program.
}
\author{
Shiyu Wang, Yan Yang, Jeff Douglas, and Steve Culpepper
}
\keyword{datasets}
