% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_rotation_data.R
\docType{data}
\name{Design_array}
\alias{Design_array}
\title{Design array}
\format{
An array of dimension N-by-J-by-L, containing each subject's item administration.
}
\source{
Spatial Rotation Learning Experiment at UIUC between Fall 2015 and Spring 2016.
}
\usage{
Design_array
}
\description{
\code{Design_array} contains item administration information at all time points in the Spatial
Rotation Learning Program.
}
\details{
The data object \code{"Design_array"} contains an array of dimension N-by-J-by-L
indicating the items assigned (1/0) to each subject at each time point.
}
\author{
Shiyu Wang, Yan Yang, Jeff Douglas, and Steve Culpepper
}
\keyword{datasets}
