% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Q_list}
\alias{Q_list}
\title{Generate a list of Q-matrices for each examinee.}
\usage{
Q_list(Q_matrix, Test_order, Test_versions)
}
\arguments{
\item{Q_matrix}{A J-by-K matrix, indicating the item-skill relationship.}

\item{Test_order}{A TxT matrix, each row is the order of item blocks for that test version.}

\item{Test_versions}{A vector of length N, containing each subject's test version.}
}
\value{
A list of length N. Each element of the list is a JxK matrix.
}
\description{
Generate a list of length N. Each element of the list is a JxK Q_matrix of all items
administered across all time points to the examinee, in the order of administration.
}
\examples{
\donttest{
Q_examinee = Q_list(Q_matrix, Test_order, Test_versions)}
}
