% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rOmega}
\alias{rOmega}
\title{Generate a random transition matrix for the first order hidden Markov model}
\usage{
rOmega(TP)
}
\arguments{
\item{TP}{A 2^K-by-2^K dichotomous matrix of indicating possible transitions under the monotonicity assumption, created with
the TPmat function}
}
\value{
A 2^K-by-2^K transition matrix, the (i,j)th element indicating the transition probability of transitioning from i-th class to j-th class.
}
\description{
Generate a random transition matrix under nondecreasing learning trajectory assumption
}
\examples{
K = ncol(Q_matrix)
TP = TPmat(K)
Omega_sim = rOmega(TP)
}
