% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simNIDA}
\alias{simNIDA}
\title{Simulate NIDA model responses (entire cube)}
\usage{
simNIDA(alphas, Svec, Gvec, Qs, test_order, Test_versions)
}
\arguments{
\item{alphas}{An N-by-K-by-T \code{array} of attribute patterns of all persons across T time points}

\item{Svec}{A length K \code{vector} of slipping probability in applying mastered skills}

\item{Gvec}{A length K \code{vector} of guessing probability in applying mastered skills}

\item{Qs}{A J-by-K-by-T  \code{cube} of Q-matrices across all item blocks}

\item{test_order}{A N_versions-by-T \code{matrix} indicating which block of items were administered to examinees with specific test version.}

\item{Test_versions}{A length N \code{vector} of the test version of each examinee}
}
\value{
An \code{array} of NIDA item responses of examinees across all time points
}
\description{
Simulate a cube of NIDA responses for all persons on items across all time points
}
\examples{
data("Spatial_Rotation")
Svec <- runif(K,.1,.3)
Gvec <- runif(K,.1,.3)
Test_versions_sim <- sample(1:5,N,replace = T)
Y_sim = simNIDA(Alphas,Svec,Gvec,Qs,test_order,Test_versions_sim)
}
