% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.hmcdm}
\alias{print.summary.hmcdm}
\alias{summary.hmcdm}
\title{Summarizing Hidden Markov Cognitive Diagnosis Model Fits}
\usage{
\method{print}{summary.hmcdm}(x, ...)

\method{summary}{hmcdm}(object, ...)
}
\arguments{
\item{x}{an object of class "\code{hmcdm.summary}".}

\item{...}{further arguments passed to or from other methods.}

\item{object}{a fitted model object of class "\code{hmcdm}".}
}
\value{
The function \code{summary.hmcdm} computes and returns a \code{list} of point estimates of model parameters and model fit measures including DIC and PPP-values.
}
\description{
\code{summary} method for class "\code{hmcdm}" or "\code{summary.hmcdm}".
}
\examples{
\donttest{
output_FOHM = hmcdm(Y_real_array,Q_matrix,"DINA_FOHM",Design_array,1000,500)
summary(output_FOHM)
}
}
\seealso{
\code{\link[=hmcdm]{hmcdm()}}
}
