% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_rotation_data.R
\docType{data}
\name{Test_order}
\alias{Test_order}
\title{Test block ordering of each test version}
\format{
A L-by-L matrix, each row is the order of item blocks for that test version.
}
\source{
Spatial Rotation Learning Experiment at UIUC between Fall 2015 and Spring 2016.
}
\usage{
Test_order
}
\description{
\code{Test_order} contains the item block ordering corresponding to each test module.
}
\details{
Each row represents the test module number and shows the order of item blocks administered to a subject with the test module.
For example, the first row is the order of item block administration (1-2-3-4-5) to subjects with test
module 1.
}
\seealso{
\code{\link{Test_versions}}
}
\author{
Shiyu Wang, Yan Yang, Jeff Douglas, and Steve Culpepper
}
\keyword{datasets}
