% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_rotation_data.R
\docType{data}
\name{Test_versions}
\alias{Test_versions}
\title{Subjects' test version}
\format{
A vector of length N, containing each subject's assigned test module.
}
\source{
Spatial Rotation Learning Experiment at UIUC between Fall 2015 and Spring 2016.
}
\usage{
Test_versions
}
\description{
\code{Test_versions} contains each subject's test module in the Spatial Rotation Learning Program.
}
\details{
The data object \code{"Test_versions"} contains a vector of length N indicating the test module assigned to each subject.
Each test module consists of multiple item blocks with different orders over L time points.
The order of item blocks corresponding to each test module is presented in the data object \code{"Test_order"}.
}
\seealso{
\code{\link{Test_order}}
}
\author{
Shiyu Wang, Yan Yang, Jeff Douglas, and Steve Culpepper
}
\keyword{datasets}
